#8590: Added custom emit logic to save packets before sending

+++ plugins/Core/storage.py
@@ -22,8 +22,7 @@ log = logging.getLogger("storage")
 __all__ = ['InMemory', 'SQLite']
 
 class AbstractStorage(object):
-    """Abstract class for storing socket.io packets before sending.
-    """
+    """Abstract class for storing socket.io packets before sending. """
 
     def __init__(self):
         self._packet_id = 0
@@ -39,21 +38,18 @@ class AbstractStorage(object):
         return itertools.count(start)
 
     def _repopulate(self):
-        """Repopulates the queue with the packets available in storage
-        """
+        """Repopulates the queue with the packets available in storage """
         pass
 
     @property
     def id(self):
-        """Returns a unique (incremental) id for storing packets
-        """
+        """Returns a unique (incremental) id for storing packets """
         self._packet_id = next(self._counter)
         return str(self._packet_id)
 
     @property
     def lastid(self):
-        """Returns the last generated packet id
-        """
+        """Returns the last generated packet id """
         return str(self._packet_id)
 
     @property
@@ -62,41 +58,54 @@ class AbstractStorage(object):
         """
         return (self.q.qsize() > 0)
 
+    def process_done(self):
+        """Marks a queue task as done """
+        try: self.q.task_done()
+        except ValueError: pass
+
     def load(self):
         """Returns the next packet in the queue
 
         If there are no packets available in the queue then None is returned
         """
-        data = None
+        packetid = chunkid = data = None
         with self.lock:
             log.debug("Fetching packet from queue")
             try:
-                packetid, subid = self.q.get(block=False)
-                log.debug("Fetching packetid %s and chunkid %s from storage", packetid, subid)
-                data = self._load(packetid, subid)
+                packetid, chunkid = self.q.get(block=False)
+                data = self.load_packet(packetid, chunkid)
             except self.q.Empty:
                 log.debug("No packets to load from queue")
                 pass
+        return packetid, chunkid, data
+
+    def load_packet(self, packetid, chunkid=None):
+        """Returns the packet data associated with the given packetid
+        and chunkid if packet is paginated.
+
+        If there is no such packet available then None is returned
+        """
+        log.debug("Fetching packetid %s and chunkid %s from storage",
+            packetid, chunkid)
+        data = self._load(packetid, chunkid)
         return data
 
-    def save(self, packetid, subid, packet):
-        """Stores the provided packet in the queue and storage
-        """
+    def save(self, packetid, chunkid, packet):
+        """Stores the provided packet in the queue and storage """
         with self.lock:
-            log.debug("Saving packetid %s and chunkid %s to queue and storage", packetid, subid)
-            self.q.put((packetid, subid), block=False)
-            self._save(packetid, subid, packet)
+            log.debug("Saving packetid %s and chunkid %s to queue and storage",
+                packetid, chunkid)
+            self.q.put((packetid, chunkid), block=False)
+            self._save(packetid, chunkid, packet)
 
     def delete(self, packetid):
-        """Removes the packet associated with provided id from the storage
-        """
+        """Removes the packet associated with provided id from the storage """
         with self.lock:
             log.debug("Deleting packetid %s from storage", packetid)
             self._delete(packetid)
 
     def clear(self):
-        """Empties the queue and storage
-        """
+        """Empties the queue and storage """
         with self.lock:
             log.debug("Clearing the queue")
             while self.q.qsize() > 0:
@@ -105,21 +114,35 @@ class AbstractStorage(object):
             log.debug("Clearing the storage")
             self._clear()
 
+    def re_add_ids(self, items):
+        """Needs to be implemented by inheriting class """
+        items_l = len(items)
+        if items_l == 0: return
+        with self.lock:
+            with self.q.mutex:
+                self.q.unfinished_tasks += items_l
+                if items_l > 1: self.q.queue.extendleft(items)
+                else: self.q.queue.appendleft(items[0])
+
     def _save(self, *args):
         """Needs to be implemented by inheriting class """
-        raise NotImplementedError("_save method has to be explicitly implemented by inheriting class")
+        raise NotImplementedError("_save method has to be explicitly "
+            "implemented by inheriting class")
 
     def _load(self, *args):
         """Needs to be implemented by inheriting class """
-        raise NotImplementedError("_load method has to be explicitly implemented by inheriting class")
+        raise NotImplementedError("_load method has to be explicitly "
+            "implemented by inheriting class")
 
     def _delete(self, *args):
         """Needs to be implemented by inheriting class """
-        raise NotImplementedError("_delete method has to be explicitly implemented by inheriting class")
+        raise NotImplementedError("_delete method has to be explicitly "
+            "implemented by inheriting class")
 
     def _clear(self):
         """Needs to be implemented by inheriting class """
-        raise NotImplementedError("_clear method has to be explicitly implemented by inheriting class")
+        raise NotImplementedError("_clear method has to be explicitly "
+            "implemented by inheriting class")
 
 class InMemory(AbstractStorage):
     """Storing socket.io packets before sending.
@@ -136,45 +159,37 @@ class InMemory(AbstractStorage):
         self.type = 'memory'
         self.packets = {}
 
-    def _save(self, packetid, subid, packet):
-        """Stores the provided packet in the dict object
-        """
-        if subid is None:
+    def _save(self, packetid, chunkid, packet):
+        """Stores the provided packet in the dict object """
+        if chunkid is None:
             self.packets[packetid] = packet
         else:
-            if packetid not in self.packets: self.packets[packetid] = {subid: packet}
-            else: self.packets[packetid][subid] = packet
+            if packetid not in self.packets:
+                self.packets[packetid] = {chunkid: packet}
+            else: self.packets[packetid][chunkid] = packet
 
-    def _load(self, packetid, subid = None):
+    def _load(self, packetid, chunkid = None):
         """Returns the packet data associated with the given packetid
         and chunkid if packet is paginated.
 
         If there is no packet available with given id then None is returned
         """
-        pckt = self.packets[packetid]
-        if subid == "PAGIN":
-            pckt = pckt[self.page_index]
-        elif type(pckt) == dict:
-            try:
-                if subid == "" or subid is None: subid = self.page_index
-                elif subid == self.page_index: subid = 1
-                else: subid = int(subid) + 1
-                if type(subid) == int and (subid + 1) == len(pckt): subid = "eof"
-                pckt = pckt[str(subid)]
-            except:
-                pass
+        pckt = None
+        if packetid in self.packets:
+            pckt = self.packets[packetid]
+        if type(pckt) == dict:
+            if chunkid in pckt: pckt = pckt[chunkid]
+            else: pckt = None
         return pckt
 
     def _delete(self, packetid):
-        """Removes the packet from the dict for provided id
-        """
+        """Removes the packet from the dict for provided id """
         if packetid.isnumeric(): packetid = int(packetid)
         if packetid in self.packets:
             del self.packets[packetid]
 
     def _clear(self):
-        """Remove all packets from the dict.
-        """
+        """Remove all packets from the dict """
         for key in self.packets:
             del self.packets[key]
 
@@ -192,18 +207,18 @@ class SQLite(AbstractStorage):
 
     def __init__(self):
         super(SQLite, self).__init__()
-        self.type = 'file'
+        self.type = 'db'
         self.db = None
-        db_path = os.getenv("AVO_ASSURE_HOME")+os.sep+'assets'+os.sep+'packets.db'
-        if os.getenv("ICE_CLEAR_STORAGE", "") and os.path.isfile(db_path):
-            os.remove(db_path)
+        db_path = os.path.join(os.getenv("AVO_ASSURE_HOME"), 'assets',
+            'packets.db')
+        rm_db = os.getenv("ICE_CLEAR_STORAGE", "") and os.path.isfile(db_path)
+        if rm_db: os.remove(db_path)
+        self.mutex = Lock()
         self.conn = sqlite3.connect(db_path, check_same_thread=False)
         self.db = self.conn.cursor()
-        Timer(5, self._compact_db).start()
-        self.db.execute("CREATE TABLE IF NOT EXISTS packets (packetid integer, subpacketid text, packet text)")
-        last_pcktid = self.db.execute("SELECT packetid from packets").fetchall()[-1:]
-        if len(last_pcktid) != 0: self._packet_id = last_pcktid[0][0] + 1
-        self.mutex = Lock()
+        self.write("CREATE TABLE IF NOT EXISTS packets (packetid integer, "
+            "chunkid text, packet text)")
+        if not rm_db: Timer(5, self._compact_db).start()
 
     def __del__(self):
         if bool(self.db):
@@ -211,15 +226,13 @@ class SQLite(AbstractStorage):
             self.conn.close()
 
     def write(self, query, *args):
-        """Performs the write operation on (Sqlite)db
-        """
+        """Performs the write operation on (Sqlite)db """
         with self.mutex:
             self.db.execute(query, args)
             self.conn.commit()
 
     def read(self, query, *args, return_cursor=False):
-        """Performs the read operation on (Sqlite)db
-        """
+        """Performs the read operation on (Sqlite)db """
         resp = None
         with self.mutex:
             cursor = self.db.execute(query, args)
@@ -229,28 +242,26 @@ class SQLite(AbstractStorage):
                 if data and len(data) > 0: resp = data[0]
         return resp
 
-    # @property
-    # def has_packet(self):
-    #     try:
-    #         return (self.read("SELECT count(packetid) FROM packets") > 0)
-    #     except Exception as e:
-    #         log.info("Error while checking for packet availability in storage. Error: %s", e)
-    #     return False
 
     def _repopulate(self):
-        """Repopulates the queue with the packets available in storage
-        """
-        pass
-
-    def _save(self, packetid, subid, packet):
-        """Stores the provided packet in the db
-        """
+        """Repopulates the queue with the packets available in storage """
         try:
-            self.write("INSERT INTO packets VALUES (?,?,?)", packetid, subid, json.dumps(packet))
+            last_pcktid = self.read("SELECT packetid from packets",
+                return_cursor=True).fetchall()[-1:]
+            if len(last_pcktid) != 0: self._packet_id = last_pcktid[0][0] + 1
+        except Exception as e:
+            log.info("Error while restoring packet queue from storage. "
+                "Error: %s", e)
+
+    def _save(self, packetid, chunkid, packet):
+        """Stores the provided packet in the db """
+        try:
+            self.write("INSERT INTO packets VALUES (?,?,?)", packetid,
+                chunkid, json.dumps(packet))
         except Exception as e:
             log.info("Error while saving packet into storage. Error: %s", e)
 
-    def _load(self, packetid, subid = None):
+    def _load(self, packetid, chunkid = None):
         """Returns the packet data associated with the given packetid
         and chunkid if packet is paginated.
 
@@ -258,27 +269,15 @@ class SQLite(AbstractStorage):
         """
         if not packetid: return None
         try:
-            subpack_count = len(self.read("SELECT subpacketid FROM packets WHERE packetid=?", packetid, return_cursor=True).fetchall())
-            if subpack_count > 1:
-                if subid == "PAGIN" or subid == "" or subid is None:
-                    subid = self.page_index
-                elif type(subid) == int and (subid + 2) == subpack_count:
-                    subid = "eof"  # +2 for start and eof
-                elif subid == self.page_index:
-                    subid = 1
-                else:
-                    subid = int(subid) + 1
-            else:
-                subid = None
-            data = self.read("SELECT packet FROM packets WHERE packetid=? and subpacketid IS ?", packetid, subid)
+            data = self.read("SELECT packet FROM packets WHERE packetid=? and"
+                " chunkid IS ?", packetid, chunkid)
             if data: return json.loads(data)
         except Exception as e:
             log.info("Error while fetching packet from storage. Error: %s", e)
         return None
 
     def _delete(self, packetid):
-        """Removes the packet from the db for provided id
-        """
+        """Removes the packet from the db for provided id """
         if packetid.isnumeric(): packetid = int(packetid)
         try:
             self.write("DELETE FROM packets WHERE packetid=?", packetid)
@@ -286,8 +285,7 @@ class SQLite(AbstractStorage):
             log.info("Error while deleting packet from storage. Error: %s", e)
 
     def _clear(self):
-        """Remove all packets from the db.
-        """
+        """Remove all packets from the db """
         try:
             self.write("DELETE FROM packets")
             # self.read("SELECT name FROM sqlite_master where name = packets")
